requirements.txt
streamlit
pytesseract
pdf2image
openai
fpdf
PyPDF2
langchain
langchain-openai
langchain-pinecone
pinecone-client
opencv-python-headless
numpy
python-dotenv
